chore: rename file FinalPresentations
